Moved leagues table to leagues list desktop component
