+++ app/client/pages/Leagues/LeaguesList/LeaguesListDesktop/index.ts
@@ -1,0 +1,3 @@
+import LeaguesListDesktop from "./LeaguesListDesktop";
+
+export default LeaguesListDesktop;
