+++ app/client/pages/Leagues/LeaguesList/index.ts
@@ -1,0 +1,3 @@
+import LeaguesList from "./LeaguesList";
+
+export default LeaguesList;
