--- app/client/pages/Leagues/LeaguesTable/index.ts
@@ -1,3 +1,0 @@
-import LeaguesTable from "./LeaguesTable";
-
-export default LeaguesTable;
